Added preparation python scripts to insert a SFM dataset into PostgreSQL | minor io performance fixes

+++ src/utils/converters.py
@@ -56,7 +56,7 @@ def npint2pgbyte(a):
     :return: binary string
     """
 
-    l = a.astype(dtype=np.int32).tolist()
+    l = a.tolist()
 
     b = array.array('B', l).tostring()
